// packages/test/generatedCode/education/users/item/educationUserItemRequestBuilderGetRequestConfiguration.ts
import {EducationUserItemRequestBuilderGetQueryParameters} from './educationUserItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface EducationUserItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: EducationUserItemRequestBuilderGetQueryParameters;
}